requirements.txt
absl-py==1.4.0
albumentations==1.3.1
astunparse==1.6.3
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
ConfigArgParse==1.5.5
contourpy==1.1.0
cupy==12.1.0
cvxopt==1.3.2
cycler==0.11.0
fastrlock==0.8.1
ffmpeg==1.4
filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.40.0
gast==0.4.0
gdown==4.7.1
google-auth==2.21.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.0
h5py==3.9.0
idna==3.4
imageio==2.19.3
imageio-ffmpeg==0.4.8
importlib-metadata==6.7.0
importlib-resources==5.12.0
jax==0.4.13
joblib==1.3.2
keras==2.12.0
kiwisolver==1.4.4
kornia==0.7.0
lazy_loader==0.2
libclang==16.0.0
lpips==0.1.4
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.2
ml-dtypes==0.2.0
networkx==3.1
numpy 
numpy-quaternion
oauthlib==3.2.2
opencv-python==4.7.0.72
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
path==16.7.1
Pillow==9.5.0
protobuf==4.23.3
PuLP==2.8.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
pypng==0.20220715.0
PySocks==1.7.1
python-dateutil==2.8.2
pytube==15.0.0
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
qudida==0.0.4
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.18.1
scikit-learn==1.4.0
scipy==1.11.0
six==1.16.0
soupsieve==2.4.1
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorboardX==2.6.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
threadpoolctl==3.2.0
tifffile==2023.4.12
torch-tb-profiler==0.4.1
tqdm==4.65.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.3.6
wrapt==1.14.1
zipp==3.15.0
